Fix coding standard
