--- src/main/java/Commands/Command.java
@@ -1,9 +1,0 @@
-package Commands;
-
-import tasklist.TaskList;
-import ui.Ui;
-import storage.Storage;
-
-public interface Command {
-    void execute(TaskList tasks, Ui ui, Storage storage);
-}
