+++ src/main/java/commands/ListCommand.java
@@ -1,0 +1,13 @@
+package commands;
+
+import storage.Storage;
+import tasklist.TaskList;
+import ui.Ui;
+
+public class ListCommand implements Command {
+
+  @Override
+  public void execute(TaskList tasks, Ui ui, Storage storage) {
+    tasks.print();
+  }
+}
