+++ src/main/java/parser/CommandParser.java
@@ -1,16 +1,16 @@
 package parser;
 
-import Commands.AddDeadlineCommand;
-import Commands.AddEventCommand;
-import Commands.AddTodoCommand;
-import Commands.Command;
-import Commands.DeleteCommand;
-import Commands.ExitCommand;
-import Commands.HelpCommand;
-import Commands.InvalidCommand;
-import Commands.ListCommand;
-import Commands.MarkCommand;
-import Commands.UnmarkCommand;
+import commands.AddDeadlineCommand;
+import commands.AddEventCommand;
+import commands.AddTodoCommand;
+import commands.Command;
+import commands.DeleteCommand;
+import commands.ExitCommand;
+import commands.HelpCommand;
+import commands.InvalidCommand;
+import commands.ListCommand;
+import commands.MarkCommand;
+import commands.UnmarkCommand;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
